+++ themes/theme-rui/src/components/Slider.styles.ts
@@ -9,7 +9,9 @@ export const Slider: ThemeComponent<'Slider'> = {
     'absolute bg-black data-[orientation=horizontal]:h-full data-[orientation=vertical]:w-full rounded-lg',
   ]),
   thumb: cva([
-    'block h-5 w-5 rounded-full border-2 border-primary bg-background transition-colors focus-visible:outline-[3px] focus-visible:outline-ring/40 data-[disabled]:cursor-not-allowed',
+    'block h-5 w-5 rounded-full border-2 border-primary bg-background transition-colors',
+    'mixin-ring-focus-visible-borderless',
+    'data-[disabled]:cursor-not-allowed',
   ]),
   output: cva('text-text-base text-sm'),
 };